Easier to handle the input format in python than C++.
The insight is that there exists an explicit formula for the cost of choosing A everytime between two indices i and j, that cost is A * (i - j), for i >= j. So the cost of halving current work amount, i, costs B and doing the same by taking steps of 1, that cost A, costs in total A * (i - i/2). We repeat this process until we know what the cost is to reduce from N to M.