Terrible implementation, could be made significantly shorter and a bit faster.
Compute diameter of each tree. The shortest maximum diameter is found by joining the centers of the trees. New diameter is the max of diameter(tree_1), diameter(tree_2) and (diameter(tree_1) / 2) + (diameter(tree_2) / 2) + 1. The "+1" representing the edge joining the trees. Also potentially adding 1 to each tree if the dieamter is odd (one path from the center will be longer than the other)